LPD-41296 Inline/rename
