--- workspaces/liferay-aicontentwizard-workspace/client-extensions/liferay-aicontentwizard-etc-spring-boot/src/main/java/com/liferay/ai/content/wizard/enums/TemplateKeys.java
@@ -1,16 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.content.wizard.enums;
-
-/**
- * @author Keven Leone
- */
-public enum TemplateKeys {
-
-	BLANK, MASTERCLASS, MINIUM, MINIUM_FULL, RAYLIFE_AP, RAYLIFE_D2C, SPEEDWELL,
-	TEAM_EXTRANET, WELCOME
-
-}
